requirements.txt
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
tensorflow==2.8.0
matplotlib==3.5.1
scipy==1.7.3
plotly==5.5.0
flask==2.1.0
joblib==1.1.0
python-dotenv==0.20.0
openai==1.3.0
streamlit==1.28.0
